MINOR: Move `Os` class to utils package and rename it to OperatingSystem
The `common` package is public and this class is
internal.

Author: Ismael Juma <ismael@juma.me.uk>

Reviewers: Jason Gustafson <jason@confluent.io>

Closes #2759 from ijuma/move-os-to-utils
